move syscalls to separate file
