+++ utils/wasm-gen/src/lib.rs
@@ -16,7 +16,7 @@
 // You should have received a copy of the GNU General Public License
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 
-use std::{collections::BTreeMap, ops::RangeInclusive};
+use std::ops::RangeInclusive;
 
 use arbitrary::Unstructured;
 use parity_wasm::{
@@ -27,6 +27,9 @@ use parity_wasm::{
     },
 };
 use wasm_smith::{InstructionKind::*, InstructionKinds, Module as ModuleSmith, SwarmConfig};
+
+mod syscalls;
+use syscalls::{sys_calls_table, SyscallsConfig};
 
 #[cfg(test)]
 mod test;
@@ -69,42 +72,6 @@ pub struct ParamRule {
     pub restricted_ratio: Ratio,
 }
 
-#[derive(Debug, Clone)]
-pub struct SyscallConfig {
-    pub alloc_param_rules: ParamRule,
-    pub free_param_rules: ParamRule,
-    pub ptr_rule: ParamRule,
-    pub memory_size_rule: ParamRule,
-    pub no_rule: ParamRule,
-}
-
-impl Default for SyscallConfig {
-    fn default() -> Self {
-        Self {
-            alloc_param_rules: ParamRule {
-                allowed_values: 0..=512,
-                restricted_ratio: (1, 100).into(),
-            },
-            free_param_rules: ParamRule {
-                allowed_values: 0..=512,
-                restricted_ratio: (1, 100).into(),
-            },
-            ptr_rule: ParamRule {
-                allowed_values: 0..=513 * 0x10000 - 1,
-                restricted_ratio: (1, 100).into(),
-            },
-            memory_size_rule: ParamRule {
-                allowed_values: 0..=0x10000,
-                restricted_ratio: (10, 100).into(),
-            },
-            no_rule: ParamRule {
-                allowed_values: 0..=0,
-                restricted_ratio: (100, 100).into(),
-            },
-        }
-    }
-}
-
 #[derive(Clone)]
 pub struct GearConfig {
     pub process_when_no_funcs: Ratio,
@@ -116,7 +83,7 @@ pub struct GearConfig {
     pub has_mem_upper_bound: Ratio,
     pub upper_bound_can_be_less_then: Ratio,
     pub sys_call_freq: Ratio,
-    pub sys_calls: SyscallConfig,
+    pub sys_calls: SyscallsConfig,
 }
 
 impl Default for GearConfig {
@@ -212,521 +179,6 @@ pub fn gen_wasm_smith_module(u: &mut Unstructured, config: &SwarmConfig) -> Modu
             return module;
         }
     }
-}
-
-struct SysCallInfo {
-    pub params: Vec<ValueType>,
-    pub results: Vec<ValueType>,
-    pub param_rules: Vec<ParamRule>,
-    pub frequency: Ratio,
-}
-
-impl SysCallInfo {
-    pub fn func_type(&self) -> FunctionType {
-        FunctionType::new(self.params.clone(), self.results.clone())
-    }
-}
-
-/// Check that all sys calls supports by backend.
-#[test]
-fn test_sys_calls_table() {
-    use gear_backend_common::{
-        mock::MockExt,
-        Environment,
-        TerminationReason,
-    };
-    use gear_backend_wasmi::WasmiEnvironment;
-    use gear_core::message::DispatchKind;
-
-    let config = GearConfig::default();
-    let table = sys_calls_table(&config);
-
-    let mut module = builder::module().function().signature().build().build().build();
-
-    for (name, info) in table {
-        let types = module.type_section_mut().unwrap().types_mut();
-        let type_no = types.len() as u32;
-        types.push(Type::Function(info.func_type()));
-
-        module = builder::from_module(module)
-            .import()
-            .module("env")
-            .external()
-            .func(type_no)
-            .field(name)
-            .build()
-            .build();
-    }
-
-    let code = module.into_bytes().unwrap();
-
-    let mut ext = MockExt::default();
-    let res = WasmiEnvironment::execute(
-        &mut ext,
-        &code,
-        Default::default(),
-        0.into(),
-        &DispatchKind::Init,
-        |_, _| -> Result<(), u32> { Ok(()) },
-    ).unwrap();
-
-    assert_eq!(res.termination_reason, TerminationReason::Success);
-}
-
-fn sys_calls_table(config: &GearConfig) -> BTreeMap<&'static str, SysCallInfo> {
-    use ValueType::*;
-    let mut res = BTreeMap::new();
-    let frequency = config.sys_call_freq;
-
-    let ptr_rule = || config.sys_calls.ptr_rule.clone();
-    let size_rule = || config.sys_calls.memory_size_rule.clone();
-    let no_rule = || config.sys_calls.no_rule.clone();
-
-    // pub fn alloc(pages: u32) -> usize;
-    res.insert(
-        "alloc",
-        SysCallInfo {
-            params: [I32].to_vec(),
-            results: [I32].to_vec(),
-            param_rules: [config.sys_calls.alloc_param_rules.clone()].to_vec(),
-            frequency,
-        },
-    );
-    // pub fn free(page: u32);
-    res.insert(
-        "free",
-        SysCallInfo {
-            params: [I32].to_vec(),
-            results: [].to_vec(),
-            param_rules: [config.sys_calls.free_param_rules.clone()].to_vec(),
-            frequency,
-        },
-    );
-
-    // pub fn gr_debug(msg_ptr: *const u8, msg_len: u32);
-    res.insert(
-        "gr_debug",
-        SysCallInfo {
-            params: [I32, I32].to_vec(),
-            results: [].to_vec(),
-            param_rules: [ptr_rule(), size_rule()].to_vec(),
-            frequency,
-        },
-    );
-    // pub fn gr_error(data: *mut u8);
-    res.insert(
-        "gr_error",
-        SysCallInfo {
-            params: [I32].to_vec(),
-            results: [].to_vec(),
-            param_rules: [ptr_rule()].to_vec(),
-            frequency,
-        },
-    );
-
-    // pub fn gr_block_height() -> u32;
-    res.insert(
-        "gr_block_height",
-        SysCallInfo {
-            params: [].to_vec(),
-            results: [I32].to_vec(),
-            param_rules: [].to_vec(),
-            frequency,
-        },
-    );
-    // pub fn gr_block_timestamp() -> u64;
-    res.insert(
-        "gr_block_timestamp",
-        SysCallInfo {
-            params: [].to_vec(),
-            results: [I64].to_vec(),
-            param_rules: [].to_vec(),
-            frequency,
-        },
-    );
-    // pub fn gr_exit(value_dest_ptr: *const u8) -> !;
-    res.insert(
-        "gr_exit",
-        SysCallInfo {
-            params: [I32].to_vec(),
-            results: [].to_vec(),
-            param_rules: [ptr_rule()].to_vec(),
-            frequency,
-        },
-    );
-    // pub fn gr_gas_available() -> u64;
-    res.insert(
-        "gr_gas_available",
-        SysCallInfo {
-            params: [].to_vec(),
-            results: [I64].to_vec(),
-            param_rules: [].to_vec(),
-            frequency,
-        },
-    );
-    // pub fn gr_program_id(val: *mut u8);
-    res.insert(
-        "gr_program_id",
-        SysCallInfo {
-            params: [I32].to_vec(),
-            results: [].to_vec(),
-            param_rules: [ptr_rule()].to_vec(),
-            frequency,
-        },
-    );
-    // pub fn gr_origin(origin_ptr: *mut u8);
-    res.insert(
-        "gr_origin",
-        SysCallInfo {
-            params: [I32].to_vec(),
-            results: [].to_vec(),
-            param_rules: [ptr_rule()].to_vec(),
-            frequency,
-        },
-    );
-    // pub fn gr_leave() -> !;
-    res.insert(
-        "gr_leave",
-        SysCallInfo {
-            params: [].to_vec(),
-            results: [].to_vec(),
-            param_rules: [].to_vec(),
-            frequency,
-        },
-    );
-    // pub fn gr_value_available(val: *mut u8);
-    res.insert(
-        "gr_value_available",
-        SysCallInfo {
-            params: [I32].to_vec(),
-            results: [].to_vec(),
-            param_rules: [ptr_rule()].to_vec(),
-            frequency,
-        },
-    );
-    // pub fn gr_wait() -> !;
-    res.insert(
-        "gr_wait",
-        SysCallInfo {
-            params: [].to_vec(),
-            results: [].to_vec(),
-            param_rules: [].to_vec(),
-            frequency,
-        },
-    );
-    // pub fn gr_wait_no_more(duration: *const u8) -> !;
-    res.insert(
-        "gr_wait_up_to",
-        SysCallInfo {
-            params: [I32].to_vec(),
-            results: [].to_vec(),
-            param_rules: [ptr_rule()].to_vec(),
-            frequency,
-        },
-    );
-    // pub fn gr_wait_for(duration: *const u8) -> !;
-    res.insert(
-        "gr_wait_for",
-        SysCallInfo {
-            params: [I32].to_vec(),
-            results: [].to_vec(),
-            param_rules: [ptr_rule()].to_vec(),
-            frequency,
-        },
-    );
-    // pub fn gr_wake(waker_id_ptr: *const u8);
-    res.insert(
-        "gr_wake",
-        SysCallInfo {
-            params: [I32].to_vec(),
-            results: [].to_vec(),
-            param_rules: [ptr_rule()].to_vec(),
-            frequency,
-        },
-    );
-
-    // pub fn gr_exit_code() -> i32;
-    res.insert(
-        "gr_exit_code",
-        SysCallInfo {
-            params: [].to_vec(),
-            results: [I32].to_vec(),
-            param_rules: [].to_vec(),
-            frequency,
-        },
-    );
-    // pub fn gr_msg_id(val: *mut u8);
-    res.insert(
-        "gr_msg_id",
-        SysCallInfo {
-            params: [I32].to_vec(),
-            results: [].to_vec(),
-            param_rules: [ptr_rule()].to_vec(),
-            frequency,
-        },
-    );
-    // pub fn gr_read(at: u32, len: u32, dest: *mut u8);
-    res.insert(
-        "gr_exit_code",
-        SysCallInfo {
-            params: [I32, I32, I32].to_vec(),
-            results: [].to_vec(),
-            param_rules: [no_rule(), size_rule(), ptr_rule()].to_vec(),
-            frequency,
-        },
-    );
-    // pub fn gr_reply(data_ptr: *const u8, data_len: u32, value_ptr: *const u8, message_id_ptr: *mut u8) -> SyscallError;
-    res.insert(
-        "gr_reply",
-        SysCallInfo {
-            params: [I32, I32, I32, I32].to_vec(),
-            results: [I32].to_vec(),
-            param_rules: [ptr_rule(), size_rule(), ptr_rule(), ptr_rule()].to_vec(),
-            frequency,
-        },
-    );
-    // pub fn gr_reply_wgas(
-    //     data_ptr: *const u8,
-    //     data_len: u32,
-    //     gas_limit: u64,
-    //     value_ptr: *const u8,
-    //     message_id_ptr: *mut u8,
-    // ) -> SyscallError;
-    res.insert(
-        "gr_reply_wgas",
-        SysCallInfo {
-            params: [I32, I32, I64, I32, I32].to_vec(),
-            results: [I32].to_vec(),
-            param_rules: [ptr_rule(), size_rule(), no_rule(), ptr_rule(), ptr_rule()].to_vec(),
-            frequency,
-        },
-    );
-    // pub fn gr_reply_commit(value_ptr: *const u8, message_id_ptr: *mut u8) -> SyscallError;
-    res.insert(
-        "gr_reply_commit",
-        SysCallInfo {
-            params: [I32, I32].to_vec(),
-            results: [I32].to_vec(),
-            param_rules: [ptr_rule(), ptr_rule()].to_vec(),
-            frequency,
-        },
-    );
-    // pub fn gr_reply_commit_wgas(
-    //     gas_limit: u64,
-    //     value_ptr: *const u8,
-    //     message_id_ptr: *mut u8,
-    // ) -> SyscallError;
-    res.insert(
-        "gr_reply_commit_wgas",
-        SysCallInfo {
-            params: [I64, I32, I32].to_vec(),
-            results: [I32].to_vec(),
-            param_rules: [no_rule(), ptr_rule(), ptr_rule()].to_vec(),
-            frequency,
-        },
-    );
-    // pub fn gr_reply_push(data_ptr: *const u8, data_len: u32) -> SyscallError;
-    res.insert(
-        "gr_reply_push",
-        SysCallInfo {
-            params: [I32, I32].to_vec(),
-            results: [I32].to_vec(),
-            param_rules: [ptr_rule(), size_rule()].to_vec(),
-            frequency,
-        },
-    );
-    // pub fn gr_reply_to(dest: *mut u8);
-    res.insert(
-        "gr_reply_to",
-        SysCallInfo {
-            params: [I32].to_vec(),
-            results: [].to_vec(),
-            param_rules: [ptr_rule()].to_vec(),
-            frequency,
-        },
-    );
-    // pub fn gr_send(
-    //     program: *const u8,
-    //     data_ptr: *const u8,
-    //     data_len: u32,
-    //     value_ptr: *const u8,
-    //     message_id_ptr: *mut u8,
-    // ) -> SyscallError;
-    res.insert(
-        "gr_send",
-        SysCallInfo {
-            params: [I32, I32, I32, I32, I32].to_vec(),
-            results: [I32].to_vec(),
-            param_rules: [ptr_rule(), ptr_rule(), size_rule(), ptr_rule(), ptr_rule()].to_vec(),
-            frequency,
-        },
-    );
-    // pub fn gr_send_wgas(
-    //     program: *const u8,
-    //     data_ptr: *const u8,
-    //     data_len: u32,
-    //     gas_limit: u64,
-    //     value_ptr: *const u8,
-    //     message_id_ptr: *mut u8,
-    // ) -> SyscallError;
-    res.insert(
-        "gr_send_wgas",
-        SysCallInfo {
-            params: [I32, I32, I32, I64, I32, I32].to_vec(),
-            results: [I32].to_vec(),
-            param_rules: [
-                ptr_rule(),
-                ptr_rule(),
-                size_rule(),
-                no_rule(),
-                ptr_rule(),
-                ptr_rule(),
-            ]
-            .to_vec(),
-            frequency,
-        },
-    );
-    // pub fn gr_send_commit(
-    //     handle: u32,
-    //     message_id_ptr: *mut u8,
-    //     program: *const u8,
-    //     value_ptr: *const u8,
-    // ) -> SyscallError;
-    res.insert(
-        "gr_send_commit",
-        SysCallInfo {
-            params: [I32, I32, I32, I32].to_vec(),
-            results: [I32].to_vec(),
-            param_rules: [no_rule(), ptr_rule(), ptr_rule(), ptr_rule()].to_vec(),
-            frequency,
-        },
-    );
-    // pub fn gr_send_commit_wgas(
-    //     handle: u32,
-    //     message_id_ptr: *mut u8,
-    //     program: *const u8,
-    //     gas_limit: u64,
-    //     value_ptr: *const u8,
-    // ) -> SyscallError;
-    res.insert(
-        "gr_send_commit_wgas",
-        SysCallInfo {
-            params: [I32, I32, I32, I64, I32].to_vec(),
-            results: [I32].to_vec(),
-            param_rules: [no_rule(), ptr_rule(), ptr_rule(), no_rule(), ptr_rule()].to_vec(),
-            frequency,
-        },
-    );
-    // pub fn gr_send_init(handle: *mut u32) -> SyscallError;
-    res.insert(
-        "gr_send_init",
-        SysCallInfo {
-            params: [I32].to_vec(),
-            results: [I32].to_vec(),
-            param_rules: [ptr_rule()].to_vec(),
-            frequency,
-        },
-    );
-    // pub fn gr_send_push(handle: u32, data_ptr: *const u8, data_len: u32) -> SyscallError;
-    res.insert(
-        "gr_send_push",
-        SysCallInfo {
-            params: [I32, I32, I32].to_vec(),
-            results: [I32].to_vec(),
-            param_rules: [no_rule(), ptr_rule(), size_rule()].to_vec(),
-            frequency,
-        },
-    );
-    // pub fn gr_size() -> u32;
-    res.insert(
-        "gr_size",
-        SysCallInfo {
-            params: [].to_vec(),
-            results: [I32].to_vec(),
-            param_rules: [].to_vec(),
-            frequency,
-        },
-    );
-    // pub fn gr_source(program: *mut u8);
-    res.insert(
-        "gr_source",
-        SysCallInfo {
-            params: [I32].to_vec(),
-            results: [].to_vec(),
-            param_rules: [ptr_rule()].to_vec(),
-            frequency,
-        },
-    );
-    // pub fn gr_value(val: *mut u8);
-    res.insert(
-        "gr_value",
-        SysCallInfo {
-            params: [I32].to_vec(),
-            results: [].to_vec(),
-            param_rules: [ptr_rule()].to_vec(),
-            frequency,
-        },
-    );
-
-    // pub fn gr_create_program(
-    //     code_hash: *const u8,
-    //     salt_ptr: *const u8,
-    //     salt_len: u32,
-    //     data_ptr: *const u8,
-    //     data_len: u32,
-    //     value_ptr: *const u8,
-    //     program_id_ptr: *mut u8,
-    // ) -> SyscallError;
-    res.insert(
-        "gr_create_program",
-        SysCallInfo {
-            params: [I32, I32, I32, I32, I32, I32, I32].to_vec(),
-            results: [].to_vec(),
-            param_rules: [
-                ptr_rule(),
-                ptr_rule(),
-                size_rule(),
-                ptr_rule(),
-                size_rule(),
-                ptr_rule(),
-                ptr_rule(),
-            ]
-            .to_vec(),
-            frequency,
-        },
-    );
-
-    // pub fn gr_create_program_wgas(
-    //     code_hash: *const u8,
-    //     salt_ptr: *const u8,
-    //     salt_len: u32,
-    //     data_ptr: *const u8,
-    //     data_len: u32,
-    //     gas_limit: u64,
-    //     value_ptr: *const u8,
-    //     program_id_ptr: *mut u8,
-    // ) -> SyscallError;
-    res.insert(
-        "gr_create_program_wgas",
-        SysCallInfo {
-            params: [I32, I32, I32, I32, I32, I64, I32, I32].to_vec(),
-            results: [].to_vec(),
-            param_rules: [
-                ptr_rule(),
-                ptr_rule(),
-                size_rule(),
-                ptr_rule(),
-                size_rule(),
-                no_rule(),
-                ptr_rule(),
-                ptr_rule(),
-            ]
-            .to_vec(),
-            frequency,
-        },
-    );
-
-    res
 }
 
 fn make_call_instructions_vec(
